Refactor server side file structure
