+++ .gitignore
@@ -1,6 +1,5 @@
 .idea/
 node_modules/
 dist/
-ssr/server/
-ssr/dist/
+/ssr/
 .env
