+++ .prettierignore
@@ -1,6 +1,6 @@
 .idea/
 node_modules/
 dist/
-ssr/
+/ssr/
 *.css
 *.md
