+++ package.json
@@ -21,8 +21,8 @@
     "lint:ts": "npx eslint --ext .tsx,.ts src/",
     "lint:style": "npx stylelint \"src/**/*.css\"",
     "lint": "npm-run-all lint:*",
-    "prettier:write": "npx prettier --write .",
-    "prettier:check": "npx prettier --check .",
+    "prettier:write": "npx prettier --write \"src/.\" \"server/.\"",
+    "prettier:check": "npx prettier --check \"src/.\" \"server/.\"",
     "test": "npx jest",
     "db:init": "until npx prisma migrate dev && npx prisma db seed --preview-feature; do echo ...; sleep 1; done",
     "db:postinstall": "export POSTGRES_HOST=localhost; docker-compose up -d postgres && npm run db:init; docker-compose down"
