+++ server/client.tsx
@@ -1,4 +1,4 @@
-import './index.css';
+import '../src/index.css';
 import React from 'react';
 import { hydrate } from 'react-dom';
 import { Provider } from 'react-redux';
