+++ server/router/paths.ts
@@ -34,7 +34,7 @@ export const API_PATH = {
   LEADERBOARD_ALL: '/leaderboard/all',
   LEADERBOARD_UPDATE: '/leaderboard',
 
-  USER_THEME_UPDATE: '/user/theme',
+  USER_THEME: '/api/user/theme',
 
   FORUM_TOPICS: '/api/forum/topics',
   FORUM_MESSAGES: '/api/forum/topic/messages',
